Update composer packages and refactor Ansible playbooks/roles

+++ app/Observers/DomainObserver.php
@@ -52,8 +52,10 @@ class DomainObserver
      */
     public function deleted(Domain $domain)
     {
-        $domain->site->provision();
-        $domain->deleteFromServer();
+        if (!$domain->isForceDeleting()) {
+            $domain->site->provision();
+            $domain->deleteFromServer();
+        }
     }
 
     /**
